+++ lib/generators/auxiliary_rails/install_generator.rb
@@ -5,7 +5,6 @@ module AuxiliaryRails
     def install
       generate 'auxiliary_rails:install_commands'
       generate 'auxiliary_rails:install_errors'
-      generate 'auxiliary_rails:install_rubocop --no-specify-gems'
     end
   end
 end